Bug 1684696. Avoid unnecessasry casts by doing math in f32. r=aosmond
Differential Revision: https://phabricator.services.mozilla.com/D100604

+++ gfx/qcms/src/transform_util.rs
@@ -366,7 +366,7 @@ fn compute_precache_pow(output: &mut [u8; PRECACHE_OUTPUT_SIZE], gamma: f32) {
     for v in 0..PRECACHE_OUTPUT_SIZE {
         //XXX: don't do integer/float conversion... and round?
         output[v] =
-            (255.0f64 * (v as f64 / PRECACHE_OUTPUT_MAX as f64).powf(gamma as f64)) as u8;
+            (255. * (v as f32 / PRECACHE_OUTPUT_MAX as f32).powf(gamma)) as u8;
     }
 }
 pub fn compute_precache_lut(output: &mut [u8; PRECACHE_OUTPUT_SIZE], table: &[u16]) {
